+++ src/gui/settingsDialog/SettingsDialog.h
@@ -40,7 +40,7 @@ class SettingsDialog : public QDialog
 {
     Q_OBJECT
 public:
-    explicit SettingsDialog(QWidget *parent = nullptr);
+    explicit SettingsDialog(QWidget *parent, const QList<CaptureModes> &captureModes);
     ~SettingsDialog() override;
 
 private: